requirements.txt
MLB-StatsAPI
pandas
tqdm
duckdb
matplotlib
